+++ src/components/NoteFeatures/Priority.jsx
@@ -1,5 +1,5 @@
 import { v4 as uuid } from "uuid";
-import { useNote } from "../../contexts/note-context";
+import { useNote } from "../../contexts";
 
 export const Priority = () => {
   const { note, setNote } = useNote();
