+++ src/contexts/index.js
@@ -1,0 +1,2 @@
+export { useNote, NoteProvider } from "./note-context";
+export { useAuth, AuthProvider } from "./auth-context";
